[xy] Move shared folder to parent folder.
