+++ mage_ai/data_cleaner/analysis/charts.py
@@ -7,9 +7,9 @@ from mage_ai.data_cleaner.analysis.constants import (
     DATA_KEY_TIME_SERIES,
     LABEL_TYPE_RANGE,
 )
-from mage_ai.data_cleaner.shared.utils import clean_series
 from mage_ai.data_cleaner.column_types.constants import ColumnType
 from mage_ai.data_cleaner.estimators.encoders import MultipleColumnLabelEncoder
+from mage_ai.data_cleaner.shared.utils import clean_series
 import math
 import numpy as np
 import pandas as pd
