+++ mage_ai/server/websocket.py
@@ -1,6 +1,6 @@
 from jupyter_client import KernelManager
-from mage_ai.data_cleaner.shared.array import find
-from mage_ai.data_cleaner.shared.hash import merge_dict
+from mage_ai.shared.array import find
+from mage_ai.shared.hash import merge_dict
 import json
 import os
 import tornado.websocket
