+++ mage_ai/shared/constants.py
