+++ mage_ai/shared/hash.py
