+++ mage_ai/tests/data_cleaner/transformer_actions/test_base.py
@@ -1,5 +1,5 @@
 from mage_ai.data_cleaner.transformer_actions.base import BaseAction
-from mage_ai.data_cleaner.shared.hash import merge_dict
+from mage_ai.shared.hash import merge_dict
 from mage_ai.tests.base_test import TestCase
 from mage_ai.tests.data_cleaner.transformer_actions.shared import TEST_ACTION
 import numpy as np
